requirements.txt
gradio==4.19.2
openai==1.12.0
python-dotenv==1.0.0
Pillow==10.2.0
numpy==1.26.4
requests==2.31.0
fastapi==0.110.0
uvicorn==0.27.1
SQLAlchemy==2.0.29
passlib[bcrypt]==1.7.4 